moving nameForm to it's own file
