+++ src/Subcomponents/NameForm.js
@@ -1,0 +1,38 @@
+import React, { Component } from 'react';
+
+class NameForm extends React.Component {
+    constructor(props) {
+      super(props);
+      this.state = {value: ''};
+      this.handleChange = this.handleChange.bind(this);
+      this.handleSubmit = this.handleSubmit.bind(this);
+    }
+  
+    handleChange(event) {
+      this.setState({value: event.target.value});
+    }
+  
+    handleSubmit(event) {
+      console.log("Message from input: " + this.state.value);
+      try {
+      window.electronMessagePort.postMessage(this.state.value);
+      } catch (e) {
+        console.log(e)
+      }
+      event.preventDefault();
+    }
+  
+    render() {
+      return (
+        <form onSubmit={this.handleSubmit}>
+          <label>
+            Name:
+            <input type="text" value={this.state.value} onChange={this.handleChange} />
+          </label>
+          <input type="submit" value="Submit" />
+        </form>
+      );
+    }
+  }
+
+  export default NameForm;
